backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==1.0.1
whitenoise==6.6.0

# Websockets and Async
channels==4.2.0
channels-redis==4.2.1
daphne==4.1.2

# Async task processing
celery==5.4.0
redis==5.0.1

# Data processing and analysis
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.6.1
gender-guesser==0.4.0
fuzzywuzzy==0.18.0
scipy>=1.9.0
statsmodels>=0.14.0
selenium==4.29.0
seaborn==0.13.2

# Text processing and sentiment analysis
spacy==3.8.4
contractions==0.1.73
nltk==3.9.1
textblob==0.17.1
vaderSentiment==3.3.2
wordcloud==1.9.4
tabulate==0.9.0

# API Documentation
drf-spectacular==0.27.1

# Testing
pytest==8.0.0
pytest-django==4.8.0
factory-boy==3.3.0
coverage==7.6.12